- fix scons_configure.py to work with compiler flag -Werror - switch on -Wmissing-noreturn and fix all it's warnings - fix some other warnings
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@1993 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/ui/ui_basic/ui_button.cc
@@ -143,7 +143,7 @@ void Basic_Button::draw(RenderTarget* dst)
 				 -1,
 				 Widget_Cache_None,
 				 0,
-				 (m_draw_caret ? m_title.length() : -1));
+				 (m_draw_caret ? m_title.length() : (uint)-1));
 		}
 
 	// draw border
